refactor(repository): Repository interface 패키지 이동 - model/repository -> presentation/repository
