+++ app/src/main/java/woowacourse/movie/presentation/presenter/SeatSelectionPresenterImpl.kt
@@ -4,7 +4,7 @@ import woowacourse.movie.data.repository.ScreeningMovieInfoRepositoryImpl
 import woowacourse.movie.data.repository.SeatRepositoryImpl
 import woowacourse.movie.domain.model.reservation.MovieTicket
 import woowacourse.movie.domain.model.reservation.ReservationInfo
-import woowacourse.movie.domain.repository.SeatRepository
+import woowacourse.movie.presentation.repository.SeatRepository
 import woowacourse.movie.presentation.contract.SeatSelectionContract
 import woowacourse.movie.presentation.uimodel.MovieTicketUiModel
